Added TextureSet class

+++ Carbon/Maths/Maths.cpp
@@ -11,14 +11,14 @@ namespace cbn
 
 	int next_power(int value, int base)
 	{
-		return std::pow(2, std::ceil(std::log(value) / log(base)));
+		return static_cast<int>(std::pow(2, std::ceil(std::log(value) / log(base))));
 	}
 	
 	//-------------------------------------------------------------------------------------
 
 	unsigned next_power(unsigned value, unsigned base)
 	{
-		return std::pow(2, std::ceil(std::log(value) / log(base)));
+		return static_cast<int>(std::pow(2, std::ceil(std::log(value) / log(base))));
 	}
 	
 	//-------------------------------------------------------------------------------------
